script config and tweaks
